# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CryoET Catalog API",
    "version": "0.1.0"
  },
  "paths": {
    "/samples": {
      "get": {
        "tags": [
          "samples"
        ],
        "summary": "List Samples",
        "description": "Paginated list of live samples (deleted_at IS NULL) with filters and\nintrinsic child-row counts (n_acquisitions/n_tomograms/n_tilt_series).\n\nFilters are registry-driven (``catalog.api.filter_fields.FIELDS``); the\nfull param set is read off ``request.query_params`` rather than declared\none-by-one. Param naming: text/existence/boolean use the field ``key``;\nrange fields use ``{key}_min`` / ``{key}_max``. See the OpenAPI-friendly\nnote: the registry is the documented contract.\n\nFilter semantics:\n  * Repeatable categorical params act as OR within a facet, AND across.\n  * Sub-entity (1:1, label, acquisition, annotation) filters use EXISTS\n    subqueries; all conditions on one sub-entity must hold on the same row.\n  * Range filters are NULL-tolerant: a row with NULL on the bound column is\n    treated as a match (so partial metadata doesn't drop the whole sample).\n  * Counts on the SELECT list are filter-INDEPENDENT total child rows.",
        "operationId": "list_samples_samples_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "sample_id",
                "project",
                "type"
              ],
              "type": "string",
              "default": "sample_id",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SampleSummary"
                  },
                  "title": "Response List Samples Samples Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/samples/{sample_id}": {
      "get": {
        "tags": [
          "samples"
        ],
        "summary": "Get Sample",
        "description": "Full sample record with typed sub-entities, acquisitions, tomograms,\nannotations, and tilt_series. 404 for missing or soft-deleted.",
        "operationId": "get_sample_samples__sample_id__get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/manage/scans/{scan_run_id}/samples": {
      "get": {
        "tags": [
          "manage"
        ],
        "summary": "Get Scan Samples",
        "description": "Per-sample outcomes for one run (optional outcome filter). 404 if unknown.",
        "operationId": "get_scan_samples_manage_scans__scan_run_id__samples_get",
        "parameters": [
          {
            "name": "scan_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Run Id"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "upserted",
                    "skipped",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanSampleOutcomeOut"
                  },
                  "title": "Response Get Scan Samples Manage Scans  Scan Run Id  Samples Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SampleSummary": {
        "properties": {
          "sample_id": {
            "type": "string",
            "title": "Sample Id"
          },
          "project": {
            "type": "string",
            "title": "Project"
          },
          "lab_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "data_source": {
            "type": "string",
            "title": "Data Source"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "cell_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cell Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count",
            "default": 0
          },
          "n_acquisitions": {
            "type": "integer",
            "title": "N Acquisitions",
            "default": 0
          },
          "n_tomograms": {
            "type": "integer",
            "title": "N Tomograms",
            "default": 0
          },
          "n_tilt_series": {
            "type": "integer",
            "title": "N Tilt Series",
            "default": 0
          },
          "thumbnail_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Path"
          }
        },
        "type": "object",
        "required": [
          "sample_id",
          "project",
          "data_source"
        ],
        "title": "SampleSummary"
      },
...
      "ScanSampleOutcomeOut": {
        "properties": {
          "sample_id": {
            "type": "string",
            "title": "Sample Id"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "sample_id",
          "outcome"
        ],
        "title": "ScanSampleOutcomeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}